Input unified! Don't set DEBUGFLAG to true yet, though [cont'd]
Input should now be fully unified, with SAI either taking
input from the command line or Discord and outputting appropriately
based on the status of process.env.DEBUG_FLAG. Hooray!

However, debug mode is still not yet fully functional, because
output has not been unified in the Commands. So, when DEBUGFLAG
is set to true, most commands break because they always expect
a Message object that they can use to msg.channel.send() with.